// devui/src/app/modules/dnb/models/constants/behaviors.constants.ts
export enum behaviors {
  copyRow = "copyRow",
  undoCopyRow = "undoCopyRow",
  copyRowGroup = "copyRowGroup",
  undoCopyRowGroup = "undoCopyRowGroup",
  copyGroup = "copyGroup",
  undoCopyGroup = "undoCopyGroup",
  copyColumn = "copyColumn",
  pasteColumn = "pasteColumn",
  removeRow = "removeRow",
  undoRemoveRow = "undoRemoveRow",
  removeGroup = "removeGroup",
  undoRemoveGroup = "undoRemoveGroup",
  addRow = "addRow",
  addGroup = "addGroup",
  movedRow = "movedRow",
  movedGroup = "movedGroup",
  addGroupRow = "addGroupRow",
  duplicateDosing = "duplicateDosing",
  newChanged = "newChanged",
}